third_party/blink/web_tests/fast/dom/incompatible-operations-expected.txt

This test checks the behavior of DOM operations when called on non-DOM or incompatible receivers with non-DOM or incompatible arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS aNode.appendChild(aDOMImplementation) threw exception TypeError: Failed to execute 'appendChild' on 'Node': parameter 1 is not of type 'Node'..
PASS aNode.appendChild('knort') threw exception TypeError: Failed to execute 'appendChild' on 'Node': parameter 1 is not of type 'Node'..
PASS aNode.appendChild(void 0) threw exception TypeError: Failed to execute 'appendChild' on 'Node': parameter 1 is not of type 'Node'..
PASS aNode.isSameNode(aDOMImplementation) threw exception TypeError: Failed to execute 'isSameNode' on 'Node': parameter 1 is not of type 'Node'..
PASS aNode.isSameNode('foo') threw exception TypeError: Failed to execute 'isSameNode' on 'Node': parameter 1 is not of type 'Node'..
PASS aNode.isSameNode(void 0) is false
PASS aNode.lookupPrefix(aDOMImplementation) is null
PASS aNode.lookupPrefix(void 0) is null
PASS aNode.cloneNode(aDOMImplementation) instanceof HTMLDivElement is true
PASS aSelect.add(aDOMImplementation, aDOMImplementation) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS aSelect.add(aDOMImplementation, anOption) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS aSelect.add(void 0, void 0) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS aSelect.add(void 0, anOption) threw exception TypeError: Failed to execute 'add' on 'HTMLSelectElement': The provided value is not of type '(HTMLOptGroupElement or HTMLOptionElement)'..
PASS aSelect.add(anOption, aDOMImplementation) is undefined.
PASS aSelect.add(anOption, void 0) is undefined.
PASS successfullyParsed is true

TEST COMPLETE
